format: cast to ushort before calling print_hex
This may seem counterintuitive because `print_hex` is implemented for
uint anyway, but this lets the compiler know that upper bits are zeroed,
as we know for sure, because it is `%hs`. This in turn allows compiler
to prune always unused switch cases. `print_hex` is static and called
only with ushort, so even if not inlined this DCE will happen.

+++ src/format.c
@@ -75,7 +75,7 @@ void pl_str_append_vasprintf_c(void *alloc, pl_str *str, const char *fmt,
             break;
         case 'h': ; // only used for %hx
             assert(c[1] == 'x');
-            len = print_hex(buf, va_arg(ap, unsigned int));
+            len = print_hex(buf, (unsigned short) va_arg(ap, unsigned int));
             c++;
             break;
         case 'u':
